% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ryan's Mood.R
\name{predict_ryan_mood}
\alias{predict_ryan_mood}
\title{Ryan's Mood}
\usage{
predict_ryan_mood(mood, time_of_day)
}
\arguments{
\item{mood}{put in a mood you think Ryan is currently in}

\item{time_of_day}{put in a time of day.}
}
\value{

}
\description{
Ryan's Mood
}
