% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ryan's Annoyed.R
\name{why_is_ryan_angry}
\alias{why_is_ryan_angry}
\title{Why is Ryan annoyed?}
\usage{
why_is_ryan_angry(annoyed, confusion)
}
\arguments{
\item{annoyed}{inserted an adjective or mood similar to annoyed}

\item{confusion}{insert a reason why Ryan would be confused.}
}
\value{

}
\description{
Why is Ryan annoyed?
}
