% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/about_me.R
\name{about_me}
\alias{about_me}
\title{Title}
\usage{
about_me(name, day, day_number)
}
\arguments{
\item{name}{insert your name here}

\item{day}{insert a day of the week}

\item{day_number}{insert the numerical day of the week (e.g., tuesday = 5)}
}
\value{

}
\description{
Title
}
